Cambios varios

+++ dominio/Alumno.cs
@@ -8,7 +8,7 @@ namespace dominio
 {
     public class Alumno
     {
-        public int idAlumno { get; set; }
+        public long idAlumno { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
         public DateTime FechaNacimiento { get; set; }
